ASACORE-1569 ProxyBusObject caches property values
Initial C++ implementation

+++ alljoyn_core/unit_test/MsgArgTest.cc
@@ -1318,7 +1318,7 @@ TEST(MsgArgTest, MaximumLengthArrays_dictionary) {
     QStatus status;
     MsgArg arg;
     for (size_t i = 0; i < ALLJOYN_MAX_ARRAY_LEN; ++i) {
-        status = max_length_adictEntry[i].Set("{xx}", i, i);
+        status = max_length_adictEntry[i].Set("{xx}", static_cast<int64_t>(i), static_cast<int64_t>(i));
         EXPECT_EQ(ER_OK, status);
     }
     status = arg.Set("a{xx}", ALLJOYN_MAX_ARRAY_LEN, max_length_adictEntry);
